+++ src/lollms_client/__init__.py
@@ -3,7 +3,7 @@
 import ascii_colors as logging
 
 # Now import the rest of the library
-from lollms_client.lollms_core import LollmsClient, ELF_COMPLETION_FORMAT
+from lollms_client.lollms_core import LollmsClient, ELF_COMPLETION_FORMAT, LollmsBindingProfile, LollmsModelProfile
 from lollms_client.lollms_types import MSG_TYPE
 from lollms_client.lollms_discussion import LollmsDiscussion, LollmsDataManager, LollmsMessage
 from lollms_client.lollms_memory import LollmsMemoryManager, MemoryConfig, FailureMemory
@@ -26,11 +26,13 @@ from lollms_client.lollms_tools_binding import LollmsToolBinding, LollmsTOOLBind
 from lollms_client.lollms_llm_binding import LollmsLLMBindingManager
 from lollms_client.lollms_bindings_utils import list_bindings, get_binding_desc
 
-__version__ = "1.18.2" # Updated version
+__version__ = "1.18.3" # Updated version
 
 # Optionally, you could define __all__ if you want to be explicit about exports
 __all__ = [
     "LollmsClient",
+    "LollmsBindingProfile",
+    "LollmsModelProfile",
     "ELF_COMPLETION_FORMAT",
     "MSG_TYPE",
     "LollmsDiscussion",